Almost done with FCFS
*avturnaround doesnt work for some reason

- Made some test processes
- not sure if average response time formula is correct but all i did was
add all the starting times of each process and averaged it
- created a timeline function that also finds start times for each
process
- timeline doesnt have spaces but i dont know how that would work
because we are dealing with floats
- made a new constructor for Process for tests

+++ Assignment 2/src/Process.java
@@ -25,7 +25,7 @@ public class Process {
 		priority = (int) (Math.random()*(4-1+1)+1);
 	}
 	
-	public float getArrivalTime() {
+	public double getArrivalTime() {
 		return arrivalTime;
 	}
 	
@@ -37,7 +37,7 @@ public class Process {
 		return priority;
 	}
 	
-	public float getRunTime() {
+	public double getRunTime() {
 		return runTime;
 	}
